+++ packages/react/src/utils/mergeProviderContexts.ts
@@ -69,6 +69,7 @@ const mergeProviderContexts = (
     disableAnimations: false,
     target: isBrowser() ? document : undefined, // eslint-disable-line no-undef
     performance: {
+      enableSanitizeCssPlugin: process.env.NODE_ENV !== 'production',
       enableStylesCaching: true,
       enableVariablesCaching: true,
     },